Organize the components within their own directories
