+++ front-end/src/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import './index.css';
-import HomeScreen from './components/HomeScreen';
+import HomeScreen from './components/HomeScreen/HomeScreen.js';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 root.render(
